Allow client middlewares to manipulate spans
Apply the context to the inner client so that middlewares already applied to it can manipulate the created span.

This allows for several use cases, including setting the span name based on higher-level http libraries like smithy4s that allow extraction of http.route. It also facilitates logging middlewares as the span ID can be properly extracted, before they ended up with the parent span ID.